refactor ReviewModal into Common/Reviews
